+++ lib/Sabberworm/CSS/Property/AtRule.php
@@ -10,4 +10,5 @@ interface AtRule {
 	public function atRuleName();
 	public function atRuleArgs();
 	public function __toString();
+	public function render($oOutputFormat = null);
 }