+++ lib/Sabberworm/CSS/Property/Selector.php
@@ -55,6 +55,13 @@ class Selector {
 	}
 
 	public function __toString() {
+		return $this->render();
+	}
+
+	public function render($oOutputFormat = null) {
+		if($oOutputFormat === null) {
+			$oOutputFormat = new \Sabberworm\CSS\OutputFormat();
+		}
 		return $this->getSelector();
 	}