+++ lib/Sabberworm/CSS/Value/String.php
@@ -19,6 +19,13 @@ class String extends PrimitiveValue {
 	}
 
 	public function __toString() {
+		return $this->render();
+	}
+
+	public function render($oOutputFormat = null) {
+		if($oOutputFormat === null) {
+			$oOutputFormat = new \Sabberworm\CSS\OutputFormat();
+		}
 		$sString = addslashes($this->sString);
 		$sString = str_replace("\n", '\A', $sString);
 		return '"' . $sString . '"';