+++ lib/Sabberworm/CSS/Value/URL.php
@@ -20,7 +20,14 @@ class URL extends PrimitiveValue {
 	}
 
 	public function __toString() {
-		return "url({$this->oURL->__toString()})";
+		return $this->render();
+	}
+
+	public function render($oOutputFormat = null) {
+		if($oOutputFormat === null) {
+			$oOutputFormat = new \Sabberworm\CSS\OutputFormat();
+		}
+		return "url({$this->oURL->render($oOutputFormat)})";
 	}
 
 }