+++ lib/Sabberworm/CSS/Value/Value.php
@@ -3,6 +3,6 @@
 namespace Sabberworm\CSS\Value;
 
 abstract class Value {
-
 	public abstract function __toString();
+	public abstract function render($oOutputFormat = null);
 }