+++ tests/Sabberworm/CSS/RuleSet/LenientParsingTest.php
@@ -20,7 +20,7 @@ class LenientParsingTest extends \PHPUnit_Framework_TestCase {
 		$sFile = dirname(__FILE__) . '/../../../files' . DIRECTORY_SEPARATOR . "-fault-tolerance.css";
 		$oParser = new Parser(file_get_contents($sFile), Settings::create()->withLenientParsing(true));
 		$oResult = $oParser->parse();
-		$this->assertSame('.test1 {}'."\n".'.test2 {hello: 2.2;hello: 2000000000000.2;}'."\n".'#test {}'."\n".'#test2 {help: none;}'."\n", $oResult->__toString());
+		$this->assertSame('.test1 {}'."\n".'.test2 {hello: 2.2;hello: 2000000000000.2;}'."\n".'#test {}'."\n".'#test2 {help: none;}'."\n", $oResult->render());
 	}
 	
 	/**
@@ -45,14 +45,14 @@ class LenientParsingTest extends \PHPUnit_Framework_TestCase {
 		$sFile = dirname(__FILE__) . '/../../../files' . DIRECTORY_SEPARATOR . "-end-token.css";
 		$oParser = new Parser(file_get_contents($sFile), Settings::create()->withLenientParsing(true));
 		$oResult = $oParser->parse();
-		$this->assertSame("", $oResult->__toString());
+		$this->assertSame("", $oResult->render());
 	}
 
 	public function testEndToken2Positive() {
 		$sFile = dirname(__FILE__) . '/../../../files' . DIRECTORY_SEPARATOR . "-end-token-2.css";
 		$oParser = new Parser(file_get_contents($sFile), Settings::create()->withLenientParsing(true));
 		$oResult = $oParser->parse();
-		$this->assertSame('#home .bg-layout {background-image: url("/bundles/main/img/bg1.png?5");}'."\n", $oResult->__toString());
+		$this->assertSame('#home .bg-layout {background-image: url("/bundles/main/img/bg1.png?5");}'."\n", $oResult->render());
 	}
 
 	public function testLocaleTrap() {
@@ -60,14 +60,14 @@ class LenientParsingTest extends \PHPUnit_Framework_TestCase {
 		$sFile = dirname(__FILE__) . '/../../../files' . DIRECTORY_SEPARATOR . "-fault-tolerance.css";
 		$oParser = new Parser(file_get_contents($sFile), Settings::create()->withLenientParsing(true));
 		$oResult = $oParser->parse();
-		$this->assertSame('.test1 {}'."\n".'.test2 {hello: 2.2;hello: 2000000000000.2;}'."\n".'#test {}'."\n".'#test2 {help: none;}'."\n", $oResult->__toString());
+		$this->assertSame('.test1 {}'."\n".'.test2 {hello: 2.2;hello: 2000000000000.2;}'."\n".'#test {}'."\n".'#test2 {help: none;}'."\n", $oResult->render());
 	}
 
 	public function testCaseInsensitivity() {
 		$sFile = dirname(__FILE__) . '/../../../files' . DIRECTORY_SEPARATOR . "case-insensitivity.css";
 		$oParser = new Parser(file_get_contents($sFile));
 		$oResult = $oParser->parse();
-		$this->assertSame('@charset "utf-8";@import url("test.css");@media screen {}#myid {case: insensitive !important;frequency: 30Hz;font-size: 1em;color: #ff0;color: hsl(40,40%,30%);font-family: Arial;}'."\n", $oResult->__toString());
+		$this->assertSame('@charset "utf-8";@import url("test.css");@media screen {}#myid {case: insensitive !important;frequency: 30Hz;font-size: 1em;color: #ff0;color: hsl(40,40%,30%);font-family: Arial;}'."\n", $oResult->render());
 	}
 
 }